Split repo hygiene truth tests
