+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+mod ci_configs;
+mod file_size;
+mod parser_enums;
+mod removed_syntax;
+mod source_truth;
