+++ tests/docs_truth/repo_hygiene/parser_enums.rs
@@ -1,0 +1,51 @@
+use super::*;
+
+#[test]
+fn parser_type_declaration_suffixes_use_owned_keyword_enum() {
+    let source = read("src/parser/declarations.rs");
+
+    for forbidden in [
+        r#"method_name == "impl""#,
+        r#"method_name == "implements""#,
+        r#"method_name == "requires""#,
+        r#"method_name == "extends""#,
+        r#"matches!(method_name.as_str(), "implements" | "requires" | "extends")"#,
+    ] {
+        assert!(
+            !source.contains(forbidden),
+            "parser type declaration suffix dispatch should use TypeDeclarationKeyword, not raw spelling checks: {forbidden}"
+        );
+    }
+    assert!(
+        source.contains("TypeDeclarationKeyword"),
+        "parser type declaration suffix dispatch should use TypeDeclarationKeyword"
+    );
+}
+
+#[test]
+fn parser_loop_control_calls_use_owned_action_enum() {
+    for path in [
+        "src/parser/expressions.rs",
+        "src/parser/expressions/suffixes.rs",
+    ] {
+        let source = read(path);
+        for forbidden in [
+            r#"name.as_str() == "done""#,
+            r#"name.as_str() == "next""#,
+            r#"match name.as_str()"#,
+            r#""done" => Expression::LoopControl"#,
+            r#""next" => Expression::LoopControl"#,
+        ] {
+            assert!(
+                !source.contains(forbidden),
+                "{path} should parse loop control calls through LoopControlAction, not raw spelling checks: {forbidden}"
+            );
+        }
+    }
+
+    let suffixes = read("src/parser/expressions/suffixes.rs");
+    assert!(
+        suffixes.contains("name.parse::<LoopControlAction>()"),
+        "parser loop-control suffix handling should parse through LoopControlAction"
+    );
+}
